исправлены падения связанные с переполнением координат при перемещении в нулевые точки.

+++ app/SML/models/types/point/point.cpp
@@ -1,6 +1,6 @@
 #include "point.h"
 
-Point::Point(int dimension)
+Point::Point(size_t dimension)
 {
     coordinates = std::vector<double>(dimension, 0);
 }
@@ -35,7 +35,7 @@ double& Point::operator[](size_t idx)
 
 double& Point::operator[](QString name)
 {
-    int idx = SML_AXISES_NAMES.getKeyByName(name);
+    size_t idx = SML_AXISES_NAMES.getKeyByName(name);
     return operator[](idx);
 }
